Add abstract Character with common functions from Bee and Player
