+++ frontend/models/game/characters/Player.php
@@ -9,56 +9,12 @@
 namespace frontend\models\game\characters;
 
 
+use frontend\models\game\characters\base\Character;
 use frontend\models\game\characters\base\interfaces\CharacterInterface;
-use frontend\models\game\GameInterface;
 use frontend\models\game\characters\interfaces\PlayerInterface;
 
-class Player implements PlayerInterface
+class Player extends Character implements PlayerInterface
 {
-
-
-    /** @var  GameInterface */
-    private $game;
-    private $lifespan;
-
-    public function __construct(GameInterface $game)
-    {
-        $this->game = $game;
-        $this->lifespan = $this->getLifespanMax();
-    }
-
-    public function getLifespan()
-    {
-        return $this->lifespan;
-    }
-
-    public function setLifespan($value)
-    {
-        if($value < $this->getLifespanMax())
-            return $this->lifespan = $value;
-        return $this->lifespan = $this->getLifespanMax();
-    }
-
-    public function beforeDead()
-    {
-        // TODO: Implement beforeDead() method.
-    }
-
-    public function beforeTakeHit()
-    {
-        // TODO: Implement beforeTakeHit() method.
-    }
-
-    public function takeHit($criticalPercent)
-    {
-        $this->setLifespan($this->getLifespan() - $this->getHitAmount($criticalPercent));
-    }
-
-    public function afterTakeHit()
-    {
-        if($this->getLifespan() <=0 )
-            $this->toDie();
-    }
 
     public function beforeHit()
     {
@@ -75,22 +31,9 @@ class Player implements PlayerInterface
         // TODO: Implement afterHit() method.
     }
 
-    public function toDie()
+    public function beforeDead()
     {
-        $this->beforeDead();
-        $this->game->getCharacterPool()->killPlayer();
-        $this->game->finish();
-    }
-
-    public function getLifespanMax()
-    {
-        return $this->game->getConfig()['maxLifespans'][$this->getType()];
-    }
-
-    public function getHitAmount($criticalPercent)
-    {
-        $amount = $this->game->getConfig()['hitAmounts'][$this->getType()];
-        return  $amount + $amount/100*$criticalPercent;
+        $this->getGame()->getCharacterPool()->killPlayer();
     }
 
     public function getType()
